requirements.txt
api_object_schema>=1.5.1
arrow>=0.6.0
capacity>=1.3.8
colorama
confetti>=2.1.0
click
flux
gossip>=2.3.1
infi.dtypes.wwn>=0.0.2
infi.dtypes.iqn>=0.3.0
logbook>=0.11.0
munch
mitba
pact>=1.0.0
python-dateutil
requests>=2.4.0
sentinels
storage_interfaces
URLObject
vintage>=0.4.0
